refactor(type_checker): move check_builtin_function into builtins/
